refactor: separate packages by domain - auth, member (#19)
