--- src/main/java/app/pigrest/auth/controller/LoginController.java
@@ -1,20 +1,0 @@
-package app.pigrest.auth.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class LoginController {
-//    private final AuthenticationManager authenticationManager;
-//
-//    public LoginController(AuthenticationManager authenticationManager) {
-//        this.authenticationManager = authenticationManager;
-//    }
-
-//    @PostMapping("/login")
-//    public ResponseEntity<Void> login(@RequestBody LoginRequest loginRequest) {
-//        Authentication authenticationRequest = UsernamePasswordAuthenticationToken.unauthenticated(loginRequest.getUsername(), loginRequest.getPassword());
-//        Authentication authenticationResponse = authenticationManager.authenticate(authenticationRequest);
-//
-//        return null;
-//    }
-}
