+++ src/main/java/app/pigrest/auth/service/AuthService.java
@@ -3,9 +3,9 @@ package app.pigrest.auth.service;
 import app.pigrest.auth.dto.request.LoginRequest;
 import app.pigrest.auth.dto.request.RegisterRequest;
 import app.pigrest.auth.model.Auth;
-import app.pigrest.auth.model.Member;
 import app.pigrest.auth.repository.AuthRepository;
-import app.pigrest.auth.repository.MemberRepository;
+import app.pigrest.member.model.Member;
+import app.pigrest.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
